Using observers

+++ src/shots.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use crate::{constants::*, player::Player};
+use crate::constants::*;
 use bevy::{
     math::bounding::{Aabb2d, IntersectsVolume as _},
     prelude::*,
@@ -25,11 +25,12 @@ impl Plugin for ShotPlugin {
 bitflags::bitflags! {
     #[derive(PartialEq, Eq, Clone, Copy)]
     struct Layer: u32 {
-        const PLAYER_SHOT = 0b0001;
-        const ENEMY_SHOT  = 0b0010;
-        const WALL        = 0b0100;
-        const ENEMY       = 0b1000;
-        const PLAYER      = 0b1000;
+        const PLAYER_SHOT = 0b00001;
+        const ENEMY_SHOT  = 0b00010;
+        const WALL        = 0b00100;
+        const ENEMY       = 0b01000;
+        const PLAYER      = 0b01000;
+        const SHIELD      = 0b10000;
     }
 }
 
@@ -67,24 +68,15 @@ impl Collider {
         }
     }
 
+    pub fn shield_layer() -> Self {
+        Self {
+            layer: Layer::SHIELD,
+            mask: Layer::PLAYER_SHOT | Layer::ENEMY_SHOT,
+        }
+    }
+
     pub fn should_collide(a: &Collider, b: &Collider) -> bool {
         a.mask.intersects(b.layer) && b.mask.intersects(a.layer)
-    }
-
-    pub fn is_player(&self) -> bool {
-        self.layer.intersects(Layer::PLAYER)
-    }
-
-    pub fn is_enemy(&self) -> bool {
-        self.layer.intersects(Layer::ENEMY)
-    }
-
-    pub fn is_player_shot(&self) -> bool {
-        self.layer.intersects(Layer::PLAYER_SHOT)
-    }
-
-    pub fn is_enemy_shot(&self) -> bool {
-        self.layer.intersects(Layer::ENEMY_SHOT)
     }
 }
 
@@ -126,9 +118,14 @@ pub fn spawn_player_shots(mut cmds: Commands, assets: Res<AssetServer>, position
         PlayerShot,
         Collider {
             layer: Layer::PLAYER_SHOT,
-            mask: Layer::ENEMY | Layer::WALL,
+            mask: Layer::ENEMY | Layer::WALL | Layer::SHIELD,
         },
-    ));
+    ))
+    .observe(on_hit_destroy);
+}
+
+pub fn on_hit_destroy(trigger: Trigger<Hit>, mut cmds: Commands) {
+    cmds.entity(trigger.entity()).despawn();
 }
 
 pub fn spawn_enemy_shots(mut cmds: Commands, sprite: Res<SpriteWithAtlas>, position: Vec2) {
@@ -139,9 +136,10 @@ pub fn spawn_enemy_shots(mut cmds: Commands, sprite: Res<SpriteWithAtlas>, posit
         EnemyShot,
         Collider {
             layer: Layer::ENEMY_SHOT,
-            mask: Layer::PLAYER | Layer::WALL,
+            mask: Layer::PLAYER | Layer::WALL | Layer::SHIELD,
         },
-    ));
+    ))
+    .observe(on_hit_destroy);
 }
 
 fn move_player_shots(mut shots: Query<&mut Transform, With<PlayerShot>>) {
@@ -159,8 +157,8 @@ fn move_enemy_shots(mut shots: Query<(&mut Transform, &mut Sprite), With<EnemySh
     }
 }
 
-#[derive(Component)]
-struct Hit(Entity);
+#[derive(Event)]
+pub struct Hit;
 
 fn check_collisions(
     mut cmds: Commands,
@@ -171,43 +169,30 @@ fn check_collisions(
         let (entity_a, transform_a, sprite_a, collider_a) = a;
         let (entity_b, transform_b, sprite_b, collider_b) = b;
 
+        if entity_a == entity_b {
+            continue;
+        }
+
         if !Collider::should_collide(collider_a, collider_b) {
             continue;
         }
         let box_a = Aabb2d::new(
             transform_a.translation.truncate(),
-            size(sprite_a, transform_a, &images) / 2.,
+            size(sprite_a, transform_a, &images) / 3.,
         );
 
         let box_b = Aabb2d::new(
             transform_b.translation.truncate(),
-            size(sprite_b, transform_b, &images) / 2.,
+            size(sprite_b, transform_b, &images) / 3.,
         );
 
         if !box_a.intersects(&box_b) {
             continue;
         }
 
-        if collider_a.is_player() && collider_b.is_enemy_shot() {
-            on_player_hit(&mut cmds, entity_a, entity_b);
-        } else if collider_a.is_enemy_shot() && collider_b.is_player() {
-            on_player_hit(&mut cmds, entity_b, entity_a);
-        } else if collider_a.is_player_shot() && collider_b.is_enemy() {
-            on_enemy_hit(&mut cmds, entity_b, entity_a);
-        } else if collider_a.is_enemy() && collider_b.is_player_shot() {
-            on_enemy_hit(&mut cmds, entity_a, entity_b);
-        }
+        cmds.trigger_targets(Hit, entity_a);
+        cmds.trigger_targets(Hit, entity_b);
     }
-}
-
-fn on_player_hit(cmds: &mut Commands, _player: Entity, shot: Entity) {
-    dbg!("player hit");
-    cmds.entity(shot).despawn();
-}
-
-fn on_enemy_hit(cmds: &mut Commands, enemy: Entity, shot: Entity) {
-    cmds.entity(enemy).despawn();
-    cmds.entity(shot).despawn();
 }
 
 fn size(sprite: &Sprite, transform: &Transform, images: &Res<Assets<Image>>) -> Vec2 {
